requirements.txt
# requirements.txt

streamlit
openai
langchain
langchain-openai
pymongo
python-dotenv
requests
streamlit-mic-recorder
pandas